Trigger back working, see details
I now think that the input PIN30 which appears to be on 5V is just a pull up float, as when I supply a fixed pulse from the signal generator, the pin can then read the pulse correctly. Therefore, the only problem left with ultrasonic sensor is that the sensor is not actually giving any output, therefore the value in timer input PIN30 is just float.
Don't know what to do from here. Everything seems to conform to the datasheet, apart from the trigger pulse being 1ms instead of 10us, but the wording on datasheet is obviously "at least".