2 features have been added: 1. Averaging values when calibrating the ultrasound. This seems to perform slightly better and more reliably than previously. More work can still be done to improve this. 2. Scan mode settings. The settings now implemented are: setting scan speed, setting left rotation point and setting right rotation point. The only setting now implemented for scan is the number of readings to take. This is because, given how speed is set up to work, it didn't make any sense. If we change the semantics of speed then we can add it as a setting.
Both of these have been tested on the mbed and they work at least to
a reasonable extent. More rigorous testing is probably best done by
someone other than me.
Another thing to note is that in order to maintain continuity
amoungst the ui for different modes, the hash key now needs to be
pressed to go to the actual reading taking state for measure mode.