Fix bug in live/servo.c
The servo returns a `uint16_t' but the value was being cast from an `int'.
This is fine for the return value but when it is sent along network
there is a size mismatch and it segfaults.

Changed the static `current_pos' value to type `uint16_t' to fix this.